requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain==0.0.350
langchain-core==0.1.23
langchain-openai==0.0.2
langchain-community==0.0.10
langsmith==0.0.87
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
tiktoken==0.5.2
aiohttp==3.9.1
pandas==2.1.1
httpx>=0.24.0,<0.25.0
tqdm==4.66.1
scikit-learn==1.3.0
rouge-score==0.1.2
nltk==3.8.1
rank-bm25==0.2.2
sentence-transformers==2.2.2
rich==13.7.0